% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{asset_linked_to_figure}
\alias{asset_linked_to_figure}
\title{Creates an object representing a relationship between a figure and
an asset.}
\usage{
asset_linked_to_figure(figure_revision, asset_revision, use_type = "indirect")
}
\arguments{
\item{figure_revision}{figure revision ID or object}

\item{asset_revision}{asset revision ID or object}

\item{use_type}{direct or indirect}
}
\value{
relationship object
}
\description{
Creates an object representing a relationship between a figure and
an asset.
}
