% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{authenticate_jwt}
\alias{authenticate_jwt}
\title{Performs JWT authentication with username and password. Saves tokens
in the GoFigr client.}
\usage{
authenticate_jwt(gf)
}
\arguments{
\item{gf}{GoFigr client}
}
\value{
NA
}
\description{
Performs JWT authentication with username and password. Saves tokens
in the GoFigr client.
}
