% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{calc_checksum}
\alias{calc_checksum}
\title{Calculates a checksum for a file}
\usage{
calc_checksum(path)
}
\arguments{
\item{path}{path to the file}
}
\value{
checksum, as a hex digest
}
\description{
Calculates a checksum for a file
}
