% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{capture}
\alias{capture}
\title{Captures output of an expression and publishes it to GoFigr. This function
is helpful if you have multiple function calls which incrementally build a
single figure.}
\usage{
capture(expr, data = NULL, env = parent.frame())
}
\arguments{
\item{expr}{expression which generates your figure}

\item{data}{input data to publish with the figure}

\item{env}{environment in which to evaluate the captured expression (default=parent.frame())}
}
\value{
result of evaluating your expression
}
\description{
Captures output of an expression and publishes it to GoFigr. This function
is helpful if you have multiple function calls which incrementally build a
single figure.
}
