% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{cat.gofigr_revision}
\alias{cat.gofigr_revision}
\title{Default cat method for GoFigr revisions.}
\usage{
cat.gofigr_revision(x, ...)
}
\arguments{
\item{x}{revision}

\item{...}{passed to cat}
}
\value{
NA
}
\description{
Default cat method for GoFigr revisions.
}
