% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{check_configured}
\alias{check_configured}
\title{Checks whether GoFigr has been correctly configured.}
\usage{
check_configured(response = warning)
}
\arguments{
\item{response}{}
}
\description{
Checks whether GoFigr has been correctly configured.
}
