% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroom.R
\name{checkbox}
\alias{checkbox}
\title{Creates a checkbox parameter for use in \code{\link{reproducible}} functions.}
\usage{
checkbox(default)
}
\arguments{
\item{default}{default logical value}
}
\value{
a \code{gf_param} object
}
\description{
Creates a checkbox parameter for use in \code{\link{reproducible}} functions.
}
