% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{data}
\name{CONFIG_PATH}
\alias{CONFIG_PATH}
\title{Default path to the config file}
\format{
An object of class \code{character} of length 1.
}
\usage{
CONFIG_PATH
}
\value{
file path
}
\description{
Default path to the config file
}
\keyword{datasets}
