% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{create.figure}
\alias{create.figure}
\title{Creates a new figure. The created figure will be blank and won't contain
any revisions.}
\usage{
create.figure(gf, analysis, name, description = NULL)
}
\arguments{
\item{gf}{GoFigr object}

\item{analysis}{analysis under which to create the figure}

\item{name}{name for the new figure}

\item{description}{description for the new figure}
}
\value{
created figure object
}
\description{
Creates a new figure. The created figure will be blank and won't contain
any revisions.
}
