% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{create_api_key}
\alias{create_api_key}
\title{Creates a new API key. This function will only succeed if using password
authentication.}
\usage{
create_api_key(gf, name)
}
\arguments{
\item{gf}{GoFigr client. Must be using password authentication.}

\item{name}{human-readable name of the API key to create, e.g. "John's laptop"}
}
\value{
response JSON. The "token" property will contain the API key if successful.
}
\description{
Creates a new API key. This function will only succeed if using password
authentication.
}
