% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{create_asset_revision}
\alias{create_asset_revision}
\title{Creates a new asset revision}
\usage{
create_asset_revision(gf, asset, metadata = list(), data = NULL)
}
\arguments{
\item{gf}{GoFigr client}

\item{asset}{asset under which to create the revision}

\item{metadata}{metadata for the revision, as a named list}

\item{data}{list of Data objects}
}
\value{
created revision object
}
\description{
Creates a new asset revision
}
