% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{create_ggsave_args}
\alias{create_ggsave_args}
\title{Creates a list of arguments for ggplot2::ggsave() with optional dimension and DPI parameters.}
\usage{
create_ggsave_args(
  filename,
  plot,
  width = NULL,
  height = NULL,
  units = "in",
  dpi = NULL
)
}
\arguments{
\item{filename}{output filename}

\item{plot}{plot object}

\item{width}{width of the output image. If NULL, not included in arguments.}

\item{height}{height of the output image. If NULL, not included in arguments.}

\item{units}{units for width and height. If NULL, not included in arguments.}

\item{dpi}{resolution of the output image. If NULL, not included in arguments.}
}
\value{
list of arguments suitable for do.call(ggplot2::ggsave, ...)
}
\description{
Creates a list of arguments for ggplot2::ggsave() with optional dimension and DPI parameters.
}
