% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision.R
\docType{data}
\name{DATA_TYPES}
\alias{DATA_TYPES}
\title{List of data types supported by GoFigr}
\format{
An object of class \code{list} of length 5.
}
\usage{
DATA_TYPES
}
\description{
List of data types supported by GoFigr
}
\keyword{datasets}
