% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{default.if.null}
\alias{default.if.null}
\title{Returns a default value if argument is null or empty}
\usage{
default.if.null(x, default)
}
\arguments{
\item{x}{argument}

\item{default}{default value if x is null, NA or ""}
}
\value{
x if not null, NA or "", or the default value
}
\description{
Returns a default value if argument is null or empty
}
