% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroom.R
\name{dropdown}
\alias{dropdown}
\title{Creates a dropdown parameter for use in \code{\link{reproducible}} functions.}
\usage{
dropdown(default, choices)
}
\arguments{
\item{default}{default value}

\item{choices}{character vector of allowed values}
}
\value{
a \code{gf_param} object
}
\description{
Creates a dropdown parameter for use in \code{\link{reproducible}} functions.
}
