% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{enable.knitr}
\alias{enable.knitr}
\title{Enables GoFigr within knitr.}
\usage{
enable.knitr(
  analysis.api.id = NULL,
  analysis.name = NULL,
  workspace = NULL,
  create.analysis = TRUE,
  analysis.description = NULL
)
}
\arguments{
\item{analysis.api.id}{Analysis API ID (if not using analysis.name)}

\item{analysis.name}{Analysis name (if not using analysis.api.id)}

\item{workspace}{parent workspace}

\item{create.analysis}{if TRUE and the analysis name does not exist, it will be created}

\item{analysis.description}{description for the analysis if it needs to be created}
}
\value{
GoFigr client
}
\description{
Enables GoFigr within knitr.
}
