% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{enable}
\alias{enable}
\title{Enables GoFigr support.}
\usage{
enable(
  analysis_api_id = NULL,
  analysis_name = NULL,
  workspace = NULL,
  create_analysis = TRUE,
  analysis_description = NULL,
  watermark = QR_WATERMARK,
  auto_publish = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{analysis_api_id}{Analysis API ID (if analysis_name is NULL)}

\item{analysis_name}{Analysis name (if analysis_api_id is NULL)}

\item{workspace}{API ID of the workspace}

\item{create_analysis}{if TRUE and analysis_name does not exist, it will be automatically created}

\item{analysis_description}{analysis description if creating a new analysis}

\item{watermark}{watermark class to use, e.g. QR_WATERMARK, LINK_WATERMARK or NO_WATERMARK}

\item{auto_publish}{will publish all plots automatically if TRUE}

\item{verbose}{whether to show verbose output}
}
\value{
named list of GoFigr options
}
\description{
Enables GoFigr support.
}
