% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{enable_knitr}
\alias{enable_knitr}
\title{Enables GoFigr within knitr.}
\usage{
enable_knitr(
  analysis_api_id = NULL,
  analysis_name = NULL,
  workspace = NULL,
  create_analysis = TRUE,
  analysis_description = NULL,
  im_options = "-density 300",
  watermark = QR_WATERMARK
)
}
\arguments{
\item{analysis_api_id}{Analysis API ID (if not using analysis_name)}

\item{analysis_name}{Analysis name (if not using analysis_api_id)}

\item{workspace}{parent workspace}

\item{create_analysis}{if TRUE and the analysis name does not exist, it will be created}

\item{analysis_description}{description for the analysis if it needs to be created}

\item{watermark}{type of watermark to use. See watermark_generator() to customize. NULL for no watermark.}
}
\value{
GoFigr client
}
\description{
Enables GoFigr within knitr.
}
