% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{encode.raw.data}
\alias{encode.raw.data}
\title{Converts a GoFigr data object into R primitives that can be converted to JSON,
performing base64 encoding of binary data.}
\usage{
encode.raw.data(data)
}
\arguments{
\item{data}{GoFigr data object}
}
\value{
encoded data object
}
\description{
Converts a GoFigr data object into R primitives that can be converted to JSON,
performing base64 encoding of binary data.
}
\examples{
data <- make.raw.data("test", "text", list(a=1), charToRaw("abcdefksjdfklsd"))
encode.raw.data(data)
}
