% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{encode_raw_data}
\alias{encode_raw_data}
\title{Converts a GoFigr data object into R primitives that can be converted to JSON,
performing base64 encoding of binary data.}
\usage{
encode_raw_data(data)
}
\arguments{
\item{data}{GoFigr data object}
}
\value{
encoded data object
}
\description{
Converts a GoFigr data object into R primitives that can be converted to JSON,
performing base64 encoding of binary data.
}
\examples{
data <- make_raw_data("test", "text", list(a=1), charToRaw("abcdefksjdfklsd"))
encode_raw_data(data)
}
