% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{find_config}
\alias{find_config}
\title{Finds the .gofigr config file in current directory or any of the
parent directories. If the file cannot be found, will also check
CONFIG_PATH.}
\usage{
find_config(start_dir = NULL)
}
\arguments{
\item{start_dir}{top-level directory where to start looking. getwd()
by default.}
}
\value{
path to .gofigr, or NULL if not found
}
\description{
Finds the .gofigr config file in current directory or any of the
parent directories. If the file cannot be found, will also check
CONFIG_PATH.
}
