% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{get.figure}
\alias{get.figure}
\title{Fetches a figure given an API ID.}
\usage{
get.figure(gf, api_id)
}
\arguments{
\item{gf}{GoFigr client}

\item{api_id}{API ID for the figure}
}
\value{
figure object
}
\description{
Fetches a figure given an API ID.
}
