% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{get.workspace}
\alias{get.workspace}
\title{Retrieves workspace details.}
\usage{
get.workspace(gf, api_id)
}
\arguments{
\item{gf}{GoFigr client}

\item{api_id}{API ID of the workspace}
}
\value{
workspace details
}
\description{
Retrieves workspace details.
}
\examples{
get.workspace("59da9bdb-2095-47a9-b414-c029f8a00e0e")
}
