% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_analysis}
\alias{get_analysis}
\title{Fetch an analysis by API ID.}
\usage{
get_analysis(gf, api_id)
}
\arguments{
\item{gf}{GoFigr client.}

\item{api_id}{Character string with the API ID of the analysis to fetch.}
}
\value{
An analysis object as returned by the API.
}
\description{
Fetch an analysis by API ID.
}
