% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_api_id}
\alias{get_api_id}
\title{Returns obj$api_id if argument is an object, or identity if it's a string.}
\usage{
get_api_id(obj)
}
\arguments{
\item{obj}{object for which to get the API ID}
}
\value{
API ID, a string
}
\description{
Returns obj$api_id if argument is an object, or identity if it's a string.
}
