% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{get_asset}
\alias{get_asset}
\title{Fetches an asset given an API ID.}
\usage{
get_asset(gf, api_id)
}
\arguments{
\item{gf}{GoFigr client}

\item{api_id}{API ID for the asset}
}
\value{
asset object
}
\description{
Fetches an asset given an API ID.
}
