% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{get_asset_revision}
\alias{get_asset_revision}
\title{Gets an asset revision given an API ID}
\usage{
get_asset_revision(gf, api_id)
}
\arguments{
\item{gf}{GoFigr client}

\item{api_id}{API ID for the revision}
}
\value{
asset revision object
}
\description{
Gets an asset revision given an API ID
}
