% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data}
\alias{get_data}
\title{Retrieves a data object. Use in conjunction with get_revision or get_asset_revision,
to retrieve the full data for a data object.}
\usage{
get_data(gf, api_id)
}
\arguments{
\item{gf}{GoFigr client}

\item{api_id}{API ID of the data object}
}
\value{
full data object
}
\description{
Retrieves a data object. Use in conjunction with get_revision or get_asset_revision,
to retrieve the full data for a data object.
}
