% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{get_execution_context}
\alias{get_execution_context}
\title{Gets the execution context: input path, chunk code, and other metadata.}
\usage{
get_execution_context()
}
\value{
named list with the execution context.
}
\description{
Gets the execution context: input path, chunk code, and other metadata.
}
\examples{
get_execution_context()
}
