% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{get_options}
\alias{get_options}
\title{Gets configured GoFigr options.}
\usage{
get_options(env = .GlobalEnv)
}
\arguments{
\item{env}{Which environment to get the options from.}
}
\value{
GoFigr options, or NULL if not set.
}
\description{
Gets configured GoFigr options.
}
