% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_components.R
\name{get_qr_png}
\alias{get_qr_png}
\title{Generates a QR code and converts it into an img element.}
\usage{
get_qr_png(url, xres = 400, yres = 400, width = 100, height = 100)
}
\arguments{
\item{url}{URL to generate the code for}

\item{xres}{QR width, in pixels}

\item{yres}{QR height, in pixels}

\item{width}{width of the HTML img element}

\item{height}{height of the HTML img element}
}
\value{
HTML string
}
\description{
Generates a QR code and converts it into an img element.
}
