% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision.R
\name{get_revision}
\alias{get_revision}
\title{Fetches a revision given an API ID.}
\usage{
get_revision(gf, api_id)
}
\arguments{
\item{gf}{GoFigr client}

\item{api_id}{API ID for the revision}
}
\value{
revision object
}
\description{
Fetches a revision given an API ID.
}
