% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision.R
\name{get_revision_url}
\alias{get_revision_url}
\title{Gets the full URL for a revision}
\usage{
get_revision_url(rev)
}
\arguments{
\item{rev}{revision object}
}
\value{
URL, a string
}
\description{
Gets the full URL for a revision
}
