% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{get_workspace}
\alias{get_workspace}
\title{Retrieve workspace details.}
\usage{
get_workspace(gf, api_id)
}
\arguments{
\item{gf}{GoFigr client created by `gofigr_client()`.}

\item{api_id}{Character string with the API ID of the workspace to fetch.}
}
\value{
A workspace object as returned by the API, including metadata and
  lists of analyses and members.
}
\description{
Retrieve workspace details.
}
