% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{gf_plot}
\alias{gf_plot}
\title{Plots and publishes an object (if supported)}
\usage{
gf_plot(...)
}
\arguments{
\item{...}{passed directly to plot}
}
\value{
result of the call to plot(...)
}
\description{
Plots and publishes an object (if supported)
}
