% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{gf_print}
\alias{gf_print}
\title{Prints and publishes an object (if supported)}
\usage{
gf_print(...)
}
\arguments{
\item{...}{passed directly to print}
}
\value{
result of the call to print(...)
}
\description{
Prints and publishes an object (if supported)
}
