% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_components.R
\name{gfContainer}
\alias{gfContainer}
\title{Generates a div container for the GoFigr widget.}
\usage{
gfContainer(..., jc = "center")
}
\arguments{
\item{...}{passed to the div}

\item{jc}{justify-content CSS value}
}
\value{
styled div element
}
\description{
Generates a div container for the GoFigr widget.
}
