% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_components.R
\name{gfPlot}
\alias{gfPlot}
\title{Defines a GoFigr plot area.}
\usage{
gfPlot(id, ...)
}
\arguments{
\item{id}{ID of this plot area}

\item{...}{same as plotOutput}
}
\value{
HTML elements
}
\description{
Defines a GoFigr plot area.
}
