% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_components.R
\name{gfPlotServer}
\alias{gfPlotServer}
\title{Creates a Shiny component to handle plotting and publishing. Has to be
paired with a gfPlot element in the UI.}
\usage{
gfPlotServer(
  id,
  expr,
  metadata = NULL,
  env = parent.frame(),
  figure_name = NULL,
  quoted = FALSE,
  base_graphics = FALSE
)
}
\arguments{
\item{id}{id of the gfPlot element}

\item{expr}{expression generating a plot}

\item{metadata}{metadata to publish with the figure. You can pass the shiny input object
to capture input values.}

\item{env}{environment in which to evaluate the expression}

\item{figure_name}{name of the figure to publish under. Inferred from figure's title if NULL.}

\item{quoted}{whether the passed expression is quoted}

\item{base_graphics}{whether the passed expression uses base graphics}
}
\value{
moduleServer
}
\description{
Creates a Shiny component to handle plotting and publishing. Has to be
paired with a gfPlot element in the UI.
}
