% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{ggwatermark}
\alias{ggwatermark}
\title{Applies a watermark to a plot object/function.}
\usage{
ggwatermark(qr, plot_obj, plot_func)
}
\arguments{
\item{qr}{pre-generated QR code, as an image}

\item{plot_obj}{plot object}

\item{plot_func}{plotting function}
}
\value{
ggplot object with the watermark applied
}
\description{
Applies a watermark to a plot object/function.
}
