% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{gofigr.fig.process}
\alias{gofigr.fig.process}
\title{Implements the GoFigr hook for knitr's fig.process.}
\usage{
gofigr.fig.process(path, options)
}
\arguments{
\item{path}{figure path}

\item{options}{chunk options}
}
\value{
updated figure path
}
\description{
Implements the GoFigr hook for knitr's fig.process.
}
