% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{gofigr.make_handler}
\alias{gofigr.make_handler}
\title{Wraps an HTTR method e.g. GET to provide relative URL resolution and
authentication}
\usage{
gofigr.make_handler(name, method)
}
\arguments{
\item{name}{method name, e.g. "GET"}

\item{method}{HTTR method, e.g. httr::GET}
}
\value{
wrapped method which takes a GoFigr client, a relative URL and
an expected HTTP status code.
}
\description{
Wraps an HTTR method e.g. GET to provide relative URL resolution and
authentication
}
\examples{
gofigr.GET <- gofigr.make_handler("GET", httr::GET)
responseJSON(gofigr.POST(gf, "api_key/",
                         body=jsonlite::toJSON(list(name=name),
                                               auto_unbox=TRUE),
                         content_type_json(),
             expected_status_code = 201))
}
