% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{gofigr_cat}
\alias{gofigr_cat}
\title{Equivalent to cat but only outputs if GoFigr client is verbose.}
\usage{
gofigr_cat(gf, content, ...)
}
\arguments{
\item{gf}{GoFigr client}

\item{content}{text to print}

\item{...}{passed to cat}
}
\value{
NA
}
\description{
Equivalent to cat but only outputs if GoFigr client is verbose.
}
