% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{gofigr_GET}
\alias{gofigr_GET}
\title{Wrapper for httr::GET that automatically handles authentication.}
\usage{
gofigr_GET(gf, url, expected_status_code = 200, ...)
}
\arguments{
\item{gf}{configured GoFigr client}

\item{url}{URL to make the request to, relative to the API URL e.g. user/}

\item{expected_status_code}{expected HTTP response code. We will throw
an exception if it differs.}

\item{...}{passed to the httr request function}
}
\value{
result of calling the underlying httr request function
}
\description{
Wrapper for httr::GET that automatically handles authentication.
}
