% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{gofigr_make_handler}
\alias{gofigr_make_handler}
\title{Wraps an HTTR method e.g. GET to provide relative URL resolution and
authentication}
\usage{
gofigr_make_handler(name, method)
}
\arguments{
\item{name}{method name, e.g. "GET"}

\item{method}{HTTR method, e.g. httr::GET}
}
\value{
wrapped method which takes a GoFigr client, a relative URL and
an expected HTTP status code.
}
\description{
Wraps an HTTR method e.g. GET to provide relative URL resolution and
authentication
}
