% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{intercept}
\alias{intercept}
\title{Wraps a plotting function (e.g. heatmap.2) so that its output is intercepted
by GoFigr.}
\usage{
intercept(plot_func, supported_classes = NULL)
}
\arguments{
\item{plot_func}{function to intercept}

\item{supported_classes}{calls will be intercepted *only if* the first argument is an \
instance of any of these classes}
}
\value{
intercepted function
}
\description{
Wraps a plotting function (e.g. heatmap.2) so that its output is intercepted
by GoFigr.
}
\examples{
heatmap.2 <- intercept(gplots::heatmap.2)
}
