% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{is.expired.token}
\alias{is.expired.token}
\title{Returns True if the response indicates an expired JWT token}
\usage{
is.expired.token(res)
}
\arguments{
\item{res}{httr response}
}
\value{
True if token expired
}
\description{
Returns True if the response indicates an expired JWT token
}
