% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{is_intercept_on}
\alias{is_intercept_on}
\title{Checks whether GoFigr intercept is on}
\usage{
is_intercept_on()
}
\value{
TRUE if intercept is on, FALSE otherwise
}
\description{
Checks whether GoFigr intercept is on
}
