% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroom.R
\name{is_serializable}
\alias{is_serializable}
\title{Check whether a value is serializable for clean room storage.}
\usage{
is_serializable(value)
}
\arguments{
\item{value}{the value to check}
}
\value{
TRUE if serializable, FALSE otherwise
}
\description{
Check whether a value is serializable for clean room storage.
}
