% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{LINK_WATERMARK}
\alias{LINK_WATERMARK}
\title{Draws a watermark with just a GoFigr link}
\usage{
LINK_WATERMARK(revision, image)
}
\arguments{
\item{revision}{GoFigr revision object for which to generate a watermark}

\item{image}{Magick image to which to add the watermark}
}
\value{
a function which you can pass to enable_knitr(watermark)
}
\description{
Draws a watermark with just a GoFigr link
}
