% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{list_sessions}
\alias{list_sessions}
\title{Lists all active GoFigr sessions.}
\usage{
list_sessions(pretty = TRUE)
}
\arguments{
\item{pretty}{if TRUE (default), returns a data frame summary;
if FALSE, returns the raw named list of session option environments}
}
\value{
data frame or named list of sessions
}
\description{
Lists all active GoFigr sessions.
}
