% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfconfig.R
\name{login.with.api.key}
\alias{login.with.api.key}
\title{Prompts the user for an API key or creates a new one}
\usage{
login.with.api.key(gf, max_attempts)
}
\arguments{
\item{gf}{Password-authenticated GoFigr client}

\item{max_attempts}{}
}
\value{
API key, either supplied by the user or newly created
}
\description{
Prompts the user for an API key or creates a new one
}
