% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfconfig.R
\name{login.with.username}
\alias{login.with.username}
\title{Prompts the user for username & password and logs into GoFigr}
\usage{
login.with.username(max_attempts)
}
\arguments{
\item{max_attempts}{}
}
\value{
GoFigr client
}
\description{
Prompts the user for username & password and logs into GoFigr
}
