% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfconfig.R
\name{login_with_api_key}
\alias{login_with_api_key}
\title{Prompts the user for an API key or interactively creates a new one.}
\usage{
login_with_api_key(gf, max_attempts)
}
\arguments{
\item{gf}{Authenticated GoFigr client created by `gofigr_client()`.}

\item{max_attempts}{Maximum number of attempts when validating a user-
supplied API key.}
}
\value{
A character string containing a valid API key, either supplied
  by the user or newly created.
}
\description{
Given an authenticated GoFigr client, this helper either accepts an
existing API key entered by the user or creates a new API key via the API.
The newly created key is associated with the authenticated user.
}
