% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make.image.data}
\alias{make.image.data}
\title{Creates a GoFigr data object storing image data}
\usage{
make.image.data(name, file.or.raw, format, is.watermarked, metadata = NULL)
}
\arguments{
\item{name}{name of this image}

\item{file.or.raw}{image data, either a file or a raw vector}

\item{format}{format, e.g. "png"}

\item{is.watermarked}{whether this file has a GoFigr watermark}

\item{metadata}{metadata associated with this image}
}
\value{
GoFigr data object
}
\description{
Creates a GoFigr data object storing image data
}
