% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make.raw.data}
\alias{make.raw.data}
\title{Creates a GoFigr data object which can be attached to revisions.}
\usage{
make.raw.data(name, type, metadata, data)
}
\arguments{
\item{name}{name of this data}

\item{type}{data type, e.g. DATA.TYPES$image}

\item{metadata}{metadata associated with this data object}

\item{data}{raw bytes}
}
\value{
data object
}
\description{
Creates a GoFigr data object which can be attached to revisions.
}
