% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make.text.data}
\alias{make.text.data}
\title{Creates a GoFigr data object to store text}
\usage{
make.text.data(name, contents, metadata = NULL)
}
\arguments{
\item{name}{name of this data object}

\item{contents}{contents, a character string}

\item{metadata}{metadata associated with this object}
}
\value{
GoFigr data object
}
\description{
Creates a GoFigr data object to store text
}
