% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_code_data}
\alias{make_code_data}
\title{Creates a GoFigr data object storing source code}
\usage{
make_code_data(
  name,
  contents_or_file,
  language,
  format = "text",
  metadata = NULL
)
}
\arguments{
\item{name}{name of this code object}

\item{contents_or_file}{contents, a character string or file object}

\item{language}{programming language, e.g. Python or R}

\item{format}{not supported at the moment; please use the default}

\item{metadata}{metadata associated with this object}
}
\value{
GoFigr data object
}
\description{
Creates a GoFigr data object storing source code
}
