% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_file_data}
\alias{make_file_data}
\title{Creates a GoFigr data object storing file data}
\usage{
make_file_data(name, file_or_raw, path = NULL, metadata = NULL)
}
\arguments{
\item{name}{name of this file}

\item{file_or_raw}{image data, either a file or a raw vector}

\item{path}{file path}

\item{metadata}{metadata associated with this file}
}
\value{
GoFigr data object
}
\description{
Creates a GoFigr data object storing file data
}
