% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_image_data}
\alias{make_image_data}
\title{Creates a GoFigr data object storing image data}
\usage{
make_image_data(name, file_or_raw, format, is_watermarked, metadata = NULL)
}
\arguments{
\item{name}{name of this image}

\item{file_or_raw}{image data, either a file or a raw vector}

\item{format}{format, e.g. "png"}

\item{is_watermarked}{whether this file has a GoFigr watermark}

\item{metadata}{metadata associated with this image}
}
\value{
GoFigr data object
}
\description{
Creates a GoFigr data object storing image data
}
