% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_table_data}
\alias{make_table_data}
\title{Creates a GoFigr data object storing data.frame/tabular data}
\usage{
make_table_data(name, frame, metadata = NULL)
}
\arguments{
\item{name}{name of this data object}

\item{frame}{data.frame}

\item{metadata}{metadata associated with this data object}
}
\value{
GoFigr data object
}
\description{
Creates a GoFigr data object storing data.frame/tabular data
}
