% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{new_asset_revision_from_file}
\alias{new_asset_revision_from_file}
\title{Creates a new asset revision from file.}
\usage{
new_asset_revision_from_file(gf, workspace_id, path)
}
\arguments{
\item{gf}{GoFigr client}

\item{workspace_id}{parent workspace in case we have to create a brand new asset}

\item{path}{path to file}
}
\value{
asset revision object
}
\description{
Creates a new asset revision from file.
}
