% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\docType{data}
\name{NO_WATERMARK}
\alias{NO_WATERMARK}
\title{Does not draw any watermarks.}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
NO_WATERMARK
}
\value{
does not return anything (NULL)
}
\description{
Does not draw any watermarks.
}
\keyword{datasets}
