% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroom.R
\name{param_to_shiny_input}
\alias{param_to_shiny_input}
\title{Build a Shiny input widget from a gf_param descriptor.}
\usage{
param_to_shiny_input(ns, name, desc)
}
\arguments{
\item{ns}{Shiny namespace function}

\item{name}{parameter name}

\item{desc}{gf_param descriptor}
}
\value{
Shiny input element, or NULL for static params
}
\description{
Build a Shiny input widget from a gf_param descriptor.
}
