% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_interactive}
\alias{plot_interactive}
\title{Replacement for plot in an interactive session outside of RStudio (e.g. terminal).
Captures the plot and publishes to GoFigr.}
\usage{
plot_interactive(..., base_func)
}
\arguments{
\item{...}{same as plot()}
}
\value{
same as plot()
}
\description{
Replacement for plot in an interactive session outside of RStudio (e.g. terminal).
Captures the plot and publishes to GoFigr.
}
