% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_knitr}
\alias{plot_knitr}
\title{Replacement for plot() in a knitr context. Captures the plot and publishes
it to GoFigr.}
\usage{
plot_knitr(..., base_func)
}
\arguments{
\item{...}{same as plot()}
}
\value{
same as plot()
}
\description{
Replacement for plot() in a knitr context. Captures the plot and publishes
it to GoFigr.
}
