% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{print.gofigr_revision}
\alias{print.gofigr_revision}
\title{Default print method for GoFigr revisions.}
\usage{
\method{print}{gofigr_revision}(x, ...)
}
\arguments{
\item{x}{revision}

\item{...}{passed to base::print}
}
\value{
NA
}
\description{
Default print method for GoFigr revisions.
}
