% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{print.gofigrdata}
\alias{print.gofigrdata}
\title{Default print representation of GoFigr data objects.}
\usage{
\method{print}{gofigrdata}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{passed to cat}
}
\value{
NA
}
\description{
Default print representation of GoFigr data objects.
}
