% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{publish_base}
\alias{publish_base}
\title{Captures output from grid graphics (ggplot2, lattice, ComplexHeatmap, etc.)
and publishes it to GoFigr.}
\usage{
publish_base(expr, ...)
}
\arguments{
\item{expr}{the expression to plot}

\item{...}{passed through to publish()}
}
\value{
GoFigr Revision object
}
\description{
Captures output from grid graphics (ggplot2, lattice, ComplexHeatmap, etc.)
and publishes it to GoFigr.
}
