% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{QR_WATERMARK}
\alias{QR_WATERMARK}
\title{Draws a watermark with a GoFigr link and a QR code}
\usage{
QR_WATERMARK(revision, image)
}
\arguments{
\item{revision}{GoFigr revision object for which to generate a watermark}

\item{image}{Magick image to which to add the watermark}
}
\value{
a function which you can pass to enable_knitr(watermark)
}
\description{
Draws a watermark with a GoFigr link and a QR code
}
