% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{read.csv2}
\alias{read.csv2}
\title{Syncs a file with the GoFigr service and stores a reference. The file
will be associated with all figures published after this call.}
\usage{
read.csv2(path, ...)
}
\arguments{
\item{path}{path to the file}

\item{...}{passed to utils::read.csv}
}
\value{
data frame
}
\description{
Syncs a file with the GoFigr service and stores a reference. The file
will be associated with all figures published after this call.
}
