% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{read_config}
\alias{read_config}
\title{Reads the GoFigr configuration, prioritizing environment variables over the
config file:}
\usage{
read_config(path = NULL)
}
\arguments{
\item{path}{path to the config file, default find_config()}
}
\value{
parsed configuration or empty list if not available
}
\description{
* GF_USERNAME or config["username"]
* GF_PASSWORD or config["password"]
* GF_API_KEY or config["api_key"]
* GF_WORKSPACE or config["workspace"]
* GF_URL or config["url"]
}
