% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{refresh_jwt}
\alias{refresh_jwt}
\title{Refreshes the JWT access token. Attempts re-authentication if refresh fails.}
\usage{
refresh_jwt(gf)
}
\arguments{
\item{gf}{GoFigr client.}
}
\value{
NA
}
\description{
Refreshes the JWT access token. Attempts re-authentication if refresh fails.
}
