% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{responseJSON}
\alias{responseJSON}
\title{Convenience function for parsing JSON from httr responses}
\usage{
responseJSON(response)
}
\arguments{
\item{response}{httr response}
}
\value{
parsed JSON
}
\description{
Convenience function for parsing JSON from httr responses
}
