% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroom.R
\name{run_reproducible_gadget}
\alias{run_reproducible_gadget}
\title{Run the Shiny gadget for interactive clean room exploration.}
\usage{
run_reproducible_gadget(
  fn,
  descriptors,
  resolved_params,
  packages,
  clean_env,
  context,
  viewer = NULL
)
}
\arguments{
\item{fn}{the user's function}

\item{descriptors}{named list of gf_param descriptors}

\item{resolved_params}{named list of default values}

\item{packages}{package names}

\item{clean_env}{the clean execution environment}

\item{context}{the clean room context}
}
\description{
Run the Shiny gadget for interactive clean room exploration.
}
