% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{server_info}
\alias{server_info}
\title{Returns cached server info from the /info/ endpoint.}
\usage{
server_info(gf)
}
\arguments{
\item{gf}{GoFigr client}
}
\value{
named list with server info, or empty list if the call fails
}
\description{
Returns cached server info from the /info/ endpoint.
}
