% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroom.R
\name{slider}
\alias{slider}
\title{Creates a slider parameter for use in \code{\link{reproducible}} functions.}
\usage{
slider(default, min = NULL, max = NULL, step = NULL)
}
\arguments{
\item{default}{default value (integer or numeric)}

\item{min}{minimum value}

\item{max}{maximum value}

\item{step}{step size}
}
\value{
a \code{gf_param} object
}
\description{
Creates a slider parameter for use in \code{\link{reproducible}} functions.
}
