% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{stack_horizontally}
\alias{stack_horizontally}
\title{Stacks images horizontally, centering them vertically.}
\usage{
stack_horizontally(images)
}
\arguments{
\item{images}{vector of images to stack}
}
\value{
composite image
}
\description{
Stacks images horizontally, centering them vertically.
}
