% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{stack_vertically}
\alias{stack_vertically}
\title{Stacks images vertically, centering them horizontally.}
\usage{
stack_vertically(images)
}
\arguments{
\item{images}{vector of images to stack}
}
\value{
composite image
}
\description{
Stacks images vertically, centering them horizontally.
}
