% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroom.R
\name{static}
\alias{static}
\title{Creates a static (read-only) parameter for use in \code{\link{reproducible}} functions.}
\usage{
static(default)
}
\arguments{
\item{default}{the value}
}
\value{
a \code{gf_param} object
}
\description{
Use this for data frames and other non-interactive values.
}
