% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{suppress}
\alias{suppress}
\title{Suppresses any automatic GoFigr publication hooks.}
\usage{
suppress(func)
}
\arguments{
\item{func}{function in which to suppress intercepts}
}
\value{
the function with GoFigr supressed
}
\description{
Suppresses any automatic GoFigr publication hooks.
}
