% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{sync_file}
\alias{sync_file}
\title{Syncs a file with the GoFigr service and stores a reference. The file
will be associated with all figures published after this call.}
\usage{
sync_file(path)
}
\arguments{
\item{path}{path to the file}
}
\value{
null
}
\description{
Syncs a file with the GoFigr service and stores a reference. The file
will be associated with all figures published after this call.
}
