% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{sync_workspace_asset}
\alias{sync_workspace_asset}
\title{Syncs a file with the GoFigr service}
\usage{
sync_workspace_asset(gf, workspace_id, path)
}
\arguments{
\item{gf}{GoFigr client}

\item{workspace_id}{parent workspace in case we have to create a brand new asset}

\item{path}{path to file}
}
\value{
asset revision object
}
\description{
\enumerate{
  \item If we haven't seen this file before, creates a new asset and a new revision
  \item If we have seen the file but haven't seen this revision, creates a new revision
  \item If we have seen this revision, returns the existing revision
  }
}
