% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroom.R
\name{text_input}
\alias{text_input}
\title{Creates a text input parameter for use in \code{\link{reproducible}} functions.}
\usage{
text_input(default)
}
\arguments{
\item{default}{default character value}
}
\value{
a \code{gf_param} object
}
\description{
Creates a text input parameter for use in \code{\link{reproducible}} functions.
}
