% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{try_base2grob}
\alias{try_base2grob}
\title{Tries to convert expression to a grob, returning it unchanged if it fails.}
\usage{
try_base2grob(expr)
}
\arguments{
\item{expr}{expression/object to convert}
}
\value{
grob if successful, expr if not
}
\description{
Tries to convert expression to a grob, returning it unchanged if it fails.
}
