% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision.R
\name{update.revision.data}
\alias{update.revision.data}
\title{Updates data associated with a figure}
\usage{
\method{update}{revision.data}(gf, revision, new.data, silent = FALSE)
}
\arguments{
\item{gf}{GoFigr client}

\item{revision}{revision or its API ID for which to update the data}

\item{new.data}{new data, as a list of GoFigrData objects (e.g. make.image.data or make.text.data)}

\item{silent}{whether to generate an activity. Internal use only.}
}
\value{
updated revision
}
\description{
Updates data associated with a figure
}
