% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision.R
\name{update_revision_data}
\alias{update_revision_data}
\title{Updates data associated with a figure}
\usage{
update_revision_data(gf, revision, new_data, silent = FALSE, assets = list())
}
\arguments{
\item{gf}{GoFigr client}

\item{revision}{revision or its API ID for which to update the data}

\item{new_data}{new data, as a list of GoFigrData objects (e.g. make_image_data or make_text_data)}

\item{silent}{whether to generate an activity. Internal use only.}

\item{assets}{list of asset revision IDs to be assocaited with this revision}
}
\value{
updated revision
}
\description{
Updates data associated with a figure
}
