% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{user.info}
\alias{user.info}
\title{Fetches user details for the currently logged in user.}
\usage{
user.info(gf)
}
\arguments{
\item{gf}{GoFigr client}
}
\value{
user details
}
\description{
Fetches user details for the currently logged in user.
}
\examples{
user.info()
}
