% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{watermark_generator}
\alias{watermark_generator}
\title{Makes a watermark generator. You can use the result with enable(watermark=...).}
\usage{
watermark_generator(
  show_qr = TRUE,
  qr_size_px = c(100, 100),
  link_size_px = c(500, 100),
  link_bg = "#ffffff",
  font_color = "#000000",
  font_size = 14,
  font = "mono"
)
}
\arguments{
\item{show_qr}{show QR code}

\item{qr_size_px}{two-element vector specifying the width, height of the QR code}

\item{link_size_px}{two-element vector specifying the width, height of the link}

\item{link_bg}{background color for the link}

\item{font_color}{font color for the link}

\item{font_size}{font size for the link}

\item{font}{font name or family, e.g. "mono"}
}
\value{
a function which you can pass to enable_knitr(watermark)
}
\description{
Makes a watermark generator. You can use the result with enable(watermark=...).
}
