% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{with_isolated_devices}
\alias{with_isolated_devices}
\title{Executes an expression while isolating any new graphics devices it creates.}
\usage{
with_isolated_devices(expr)
}
\arguments{
\item{expr}{The R expression to evaluate.}
}
\value{
result of evaluating expr
}
\description{
Executes an expression while isolating any new graphics devices it creates.
}
