% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{check_configured}
\alias{check_configured}
\title{Checks whether GoFigr has been correctly configured.}
\usage{
check_configured(response = warning)
}
\arguments{
\item{response}{function to use to show the warning message if not
configured. Default: warning.}
}
\value{
TRUE if configured, FALSE otherwise
}
\description{
Checks whether GoFigr has been correctly configured.
}
