% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{create_analysis}
\alias{create_analysis}
\title{Creates a new analysis}
\usage{
create_analysis(gf, name, description = NULL, workspace = NULL)
}
\arguments{
\item{gf}{GoFigr client}

\item{name}{analysis name}

\item{description}{analysis description}

\item{workspace}{analysis will be created under this workspace. Can be a workspace object or an API ID.}
}
\value{
created analysis
}
\description{
Creates a new analysis
}
