% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{create_asset}
\alias{create_asset}
\title{Creates a new asset}
\usage{
create_asset(gf, workspace, name, description = NULL)
}
\arguments{
\item{gf}{GoFigr client}

\item{workspace}{parent workspace}

\item{name}{name of the asset}

\item{description}{description of the asset}
}
\value{
asset object
}
\description{
Creates a new asset
}
