% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision.R
\name{create_revision}
\alias{create_revision}
\title{Creates a new revision}
\usage{
create_revision(gf, figure, metadata = list(), data = NULL)
}
\arguments{
\item{gf}{GoFigr client}

\item{figure}{figure under which to create the revision}

\item{metadata}{metadata for the revision, as a named list}

\item{data}{list of Data objects}
}
\value{
created revision object
}
\description{
Creates a new revision
}
