% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{create_workspace}
\alias{create_workspace}
\title{Creates a new workspace}
\usage{
create_workspace(gf, name, description = NULL)
}
\arguments{
\item{gf}{GoFigr client}

\item{name}{workspace name}

\item{description}{workspace description}
}
\value{
created workspace object
}
\description{
Creates a new workspace
}
