% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{delete_analysis}
\alias{delete_analysis}
\title{Deletes an analysis given an API ID.}
\usage{
delete_analysis(gf, api_id)
}
\arguments{
\item{gf}{GoFigr client}

\item{api_id}{API ID for the analysis}
}
\value{
server response
}
\description{
Deletes an analysis given an API ID.
}
