% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{enable}
\alias{enable}
\title{Enables GoFigr support.}
\usage{
enable(
  auto_publish = FALSE,
  analysis_api_id = NULL,
  analysis_name = NULL,
  workspace = NULL,
  create_analysis = TRUE,
  analysis_description = NULL,
  watermark = QR_WATERMARK,
  verbose = FALSE,
  debug = FALSE,
  show = "watermark"
)
}
\arguments{
\item{auto_publish}{will publish all plots automatically if TRUE. Note
that setting this option will re-assign plot() and print() in the global environment.}

\item{analysis_api_id}{Analysis API ID (if analysis_name is NULL)}

\item{analysis_name}{Analysis name (if analysis_api_id is NULL)}

\item{workspace}{API ID of the workspace}

\item{create_analysis}{if TRUE and analysis_name does not exist, it will be automatically created}

\item{analysis_description}{analysis description if creating a new analysis}

\item{watermark}{watermark class to use, e.g. QR_WATERMARK, LINK_WATERMARK or NO_WATERMARK}

\item{verbose}{whether to show verbose output}

\item{debug}{whether to show debugging information}

\item{show}{which figure to display in the document: original, watermark, or hide. Note that this setting \
only affects the display and doesn't change what gets published: e.g. even if you choose to display \
the original figure, the watermarked version will still be published to GoFigr.}
}
\value{
named list of GoFigr options
}
\description{
Enables GoFigr support.
}
