% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{find_analysis}
\alias{find_analysis}
\title{Finds an analysis by name, optionally creating one if it doesn't exist.}
\usage{
find_analysis(gf, name, description = NULL, workspace = NULL, create = FALSE)
}
\arguments{
\item{gf}{GoFigr client}

\item{name}{name of the analysis to find}

\item{description}{description of the analysis if it needs to be created}

\item{workspace}{parent workspace}

\item{create}{if TRUE and the analysis doesn't exist, it will be created; throws an error otherwise.}
}
\value{
analysis object
}
\description{
Finds an analysis by name, optionally creating one if it doesn't exist.
}
