% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{find_asset_by_name}
\alias{find_asset_by_name}
\title{Finds an asset by name}
\usage{
find_asset_by_name(gf, name)
}
\arguments{
\item{gf}{GoFigr client}

\item{name}{name of the asset to search for}
}
\value{
list of matching assets, or an empty list if none found
}
\description{
Finds an asset by name
}
