% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset.R
\name{find_asset_revision_by_hash}
\alias{find_asset_revision_by_hash}
\title{Finds all asset revisions with a matching hash digest}
\usage{
find_asset_revision_by_hash(gf, digest, hash_type = "blake3")
}
\arguments{
\item{gf}{GoFigr client}

\item{digest}{hex digest string}

\item{hash_type}{digest type}
}
\value{
list of asset revisions, or empty list
}
\description{
Finds all asset revisions with a matching hash digest
}
