% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{find_figure}
\alias{find_figure}
\title{Finds a figure by name}
\usage{
find_figure(gf, analysis, name, description = NULL, create = FALSE)
}
\arguments{
\item{gf}{GoFigr client}

\item{analysis}{parent analysis}

\item{name}{name of the figure to find}

\item{description}{description of the figure if it needs to be created}

\item{create}{if TRUE and the figure doesn't exist, it will be created; throws an error otherwise.}
}
\value{
figure object
}
\description{
Finds a figure by name
}
