% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{find_workspace}
\alias{find_workspace}
\title{Finds a workspace by name, optionally creating it if doesn't exist}
\usage{
find_workspace(gf, name, description = NULL, create = FALSE)
}
\arguments{
\item{gf}{GoFigr client}

\item{name}{name of the workspace to find}

\item{description}{workspace description if creating a new one}

\item{create}{if TRUE and workspace is not found, it will be created}
}
\value{
workspace if found; throws an error if not.
}
\description{
Finds a workspace by name, optionally creating it if doesn't exist
}
