% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_analysis}
\alias{get_analysis}
\title{Fetches an analysis given an API ID.}
\usage{
get_analysis(gf, api_id)
}
\arguments{
\item{gf}{GoFigr client}

\item{api_id}{API ID for the analysis}
}
\value{
analysis object
}
\description{
Fetches an analysis given an API ID.
}
