% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{get_workspace}
\alias{get_workspace}
\title{Retrieves workspace details.}
\usage{
get_workspace(gf, api_id)
}
\arguments{
\item{gf}{GoFigr client}

\item{api_id}{API ID of the workspace}
}
\value{
workspace details
}
\description{
Retrieves workspace details.
}
