% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfconfig.R
\name{gfconfig}
\alias{gfconfig}
\title{Configures gofigr for use on this machine. Saves configuration
to ~/.gofigr.}
\usage{
gfconfig(max_attempts = 3)
}
\arguments{
\item{max_attempts}{Maximum number of password attempts}
}
\value{
No return value
}
\description{
Configures gofigr for use on this machine. Saves configuration
to ~/.gofigr.
}
