% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{infer_workspace}
\alias{infer_workspace}
\title{Returns the argument if a valid workspace is passed, or the default
workspace from the GoFigr client otherwise. Throws an error if both
are NULL.}
\usage{
infer_workspace(gf, workspace)
}
\arguments{
\item{gf}{GoFigr client}

\item{workspace}{workspace or NULL}
}
\value{
workspace object
}
\description{
Returns the argument if a valid workspace is passed, or the default
workspace from the GoFigr client otherwise. Throws an error if both
are NULL.
}
