% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{intercept}
\alias{intercept}
\title{Wraps a plotting function (e.g. heatmap.2) so that its output is intercepted
by GoFigr.}
\usage{
intercept(
  plot_func,
  supported_classes = NULL,
  force = FALSE,
  data = NULL,
  figure_name = NULL
)
}
\arguments{
\item{plot_func}{function to intercept}

\item{supported_classes}{calls will be intercepted *only if* the first argument is an \
instance of any of these classes}

\item{force}{force intercept even if auto_publish is off}

\item{data}{data to publish with the figure}

\item{figure_name}{name of the figure, if different from default}
}
\value{
intercepted function
}
\description{
Wraps a plotting function (e.g. heatmap.2) so that its output is intercepted
by GoFigr.
}
\examples{
heatmap.2 <- intercept(gplots::heatmap.2)
}
