% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{list_workspaces}
\alias{list_workspaces}
\title{List all workspaces available to the user.}
\usage{
list_workspaces(gf)
}
\arguments{
\item{gf}{GoFigr client}
}
\value{
List of workspaces
}
\description{
List all workspaces available to the user.
}
