% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfconfig.R
\name{login_with_api_key}
\alias{login_with_api_key}
\title{Prompts the user for an API key or creates a new one}
\usage{
login_with_api_key(gf, max_attempts)
}
\arguments{
\item{gf}{Password-authenticated GoFigr client}

\item{max_attempts}{Maximum number of login attempts before giving up}
}
\value{
API key, either supplied by the user or newly created
}
\description{
Prompts the user for an API key or creates a new one
}
