% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfconfig.R
\name{login_with_username}
\alias{login_with_username}
\title{Prompts the user for username & password and logs into GoFigr}
\usage{
login_with_username(max_attempts)
}
\arguments{
\item{max_attempts}{maximum number of login attempts before giving up}
}
\value{
GoFigr client
}
\description{
Prompts the user for username & password and logs into GoFigr
}
