% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{plot_knitr}
\alias{plot_knitr}
\title{Replacement for plot() in a knitr context. Captures the plot and publishes
it to GoFigr.}
\usage{
plot_knitr(..., base_func)
}
\arguments{
\item{...}{same as plot()}

\item{base_func}{the function whose output we are trying to capture}
}
\value{
same as plot()
}
\description{
Replacement for plot() in a knitr context. Captures the plot and publishes
it to GoFigr.
}
