% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{plot_rstudio}
\alias{plot_rstudio}
\title{Replacement for plot within RStudio. Captures the plot and publishes to GoFigr.}
\usage{
plot_rstudio(..., base_func)
}
\arguments{
\item{...}{same as plot()}

\item{base_func}{the function whose output we are trying to capture}
}
\value{
same as plot()
}
\description{
Replacement for plot within RStudio. Captures the plot and publishes to GoFigr.
}
