% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{plot_script}
\alias{plot_script}
\title{Replacement for plot in a script. Captures the plot and publishes to GoFigr.}
\usage{
plot_script(..., base_func)
}
\arguments{
\item{...}{same as plot()}

\item{base_func}{the function whose output we are trying to capture}
}
\value{
same as plot()
}
\description{
Replacement for plot in a script. Captures the plot and publishes to GoFigr.
}
