% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.gofigr}
\alias{print.gofigr}
\title{Default print method for a GoFigr client.}
\usage{
\method{print}{gofigr}(x, ...)
}
\arguments{
\item{x}{GoFigr client}

\item{...}{passed to cat}
}
\value{
NA
}
\description{
Default print method for a GoFigr client.
}
