% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfconfig.R
\name{read_prompt}
\alias{read_prompt}
\title{Reads a prompt from stdin and performs optional validation}
\usage{
read_prompt(prompt, validate = NULL, attempt = 1, max_attempts = 2)
}
\arguments{
\item{prompt}{prompt, e.g. "Enter username: "}

\item{validate}{function input => transformed input if input is valid or, error}

\item{attempt}{current attempt at getting a valid input}

\item{max_attempts}{maximum number of attempts}
}
\value{
input, or result of validate(input) if validate is supplied
}
\description{
Reads a prompt from stdin and performs optional validation
}
