% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{response_to_JSON}
\alias{response_to_JSON}
\title{Convenience function for parsing JSON from httr responses}
\usage{
response_to_JSON(response)
}
\arguments{
\item{response}{httr response}
}
\value{
parsed JSON
}
\description{
Convenience function for parsing JSON from httr responses
}
