% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{set_options}
\alias{set_options}
\title{Sets GoFigr options.}
\usage{
set_options(options)
}
\arguments{
\item{options}{New options that will replace existing options.}
}
\value{
NA
}
\description{
Sets GoFigr options.
}
